pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI Agent guided project part of Boot.dev's curriculum."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "pytest>=8.4.2",
    "python-dotenv==1.1.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = "src"
